requirements.txt
transformers==4.49.0
peft
datasets
pyyaml